@@ -1,9 +1,11 @@
-import { Component, OnInit, ViewChild, ElementRef } from '@angular/core';
+import { Component, OnInit, ViewChild, ElementRef, OnDestroy } from '@angular/core';
 import { FormGroup, FormBuilder, Validators } from '@angular/forms';
 import { ActivatedRoute, ParamMap, Router } from '@angular/router';
 import { select, Store } from '@ngrx/store';
-import { Observable, Subscription } from 'rxjs/Rx';
+import { of as observableOf } from 'rxjs';
 import { switchMap } from 'rxjs/operators';
+import { Subject } from 'rxjs/Subject';
+import { takeUntil } from 'rxjs/operators';
 import { find, get, pull } from 'lodash';
 
 import { Blog } from '../../model/blog';
@@ -15,7 +17,8 @@ import * as BlogActions from '../../actions/blog.actions';
   templateUrl: './edit-blog.component.html',
   styleUrls: ['./edit-blog.component.scss'],
 })
-export class EditBlogComponent implements OnInit {
+export class EditBlogComponent implements OnInit, OnDestroy {
+  destroy$: Subject<void> = new Subject<void>();
   @ViewChild('tagInput')
   tagInputRef: ElementRef;
 
@@ -23,9 +26,7 @@ export class EditBlogComponent implements OnInit {
   form: FormGroup;
   blogId: string;
   editMode: boolean;
-  showPreview: boolean = true;
-  initBlogSubscription: Subscription;
-  errorMessageSubscription: Subscription;
+  showPreview = true;
   tags: string[] = [];
 
   constructor(
@@ -36,7 +37,7 @@ export class EditBlogComponent implements OnInit {
   ) {}
 
   ngOnInit(): void {
-    this.initBlogSubscription = this.route.paramMap
+    this.route.paramMap
       .pipe(
         switchMap((params: ParamMap) => {
           this.blogId = params.get('id');
@@ -45,9 +46,10 @@ export class EditBlogComponent implements OnInit {
             this.store.dispatch(new BlogActions.LoadOneBlog(params.get('id')));
             return this.store.pipe(select(fromBlog.getSelectedBlog));
           } else {
-            return Observable.of(undefined);
+            return observableOf(undefined);
           }
-        })
+        }),
+        takeUntil(this.destroy$)
       )
       .subscribe((blog: Blog) => {
         this.blog = blog;
@@ -98,9 +100,10 @@ export class EditBlogComponent implements OnInit {
     });
   }
 
-  save(): void {
+  save(isDraft: boolean): void {
     const formValue = this.form.getRawValue();
-    let blog = {
+    const blog = {
+      isDraft,
       title: formValue.title,
       content: formValue.content,
       lastModified: new Date().valueOf(),
@@ -109,13 +112,14 @@ export class EditBlogComponent implements OnInit {
     if (this.editMode) {
       this.store.dispatch(new BlogActions.EditBlog({ id: this.blogId, blog }));
     } else {
-      this.store.dispatch(
-        new BlogActions.AddBlog({ ...blog, createdDate: new Date().valueOf() })
-      );
+      this.store.dispatch(new BlogActions.AddBlog({ ...blog, createdDate: new Date().valueOf() }));
     }
-    this.errorMessageSubscription = this.store
-      .pipe(select(fromBlog.getErrorMessage))
-      .subscribe(errorMessage => {
+    this.store
+      .pipe(
+        select(fromBlog.getErrorMessage),
+        takeUntil(this.destroy$)
+      )
+      .subscribe((errorMessage) => {
         if (errorMessage === 'Success') {
           this.router.navigate(['/blog']);
         }
@@ -131,11 +135,7 @@ export class EditBlogComponent implements OnInit {
   }
 
   ngOnDestroy(): void {
-    if (this.initBlogSubscription) {
-      this.initBlogSubscription.unsubscribe();
-    }
-    if (this.errorMessageSubscription) {
-      this.errorMessageSubscription.unsubscribe();
-    }
+    this.destroy$.next();
+    this.destroy$.complete();
   }
 }
