@@ -8,18 +8,28 @@ export class Blog {
   summary: string;
   tags: string[];
   lastModified: string;
+  isDraft: boolean;
 
   constructor(blog: any) {
     this.id = blog.id;
     this.title = blog.title;
     this.content = blog.content;
     this.summary = blog.summary;
-    this.createdDate = dayjs(blog.createdDate).format('MMM DD, YYYY');
-    this.lastModified = dayjs(blog.lastModified).format('MMM DD, YYYY');
+    this.createdDate = blog.createdDate;
+    this.lastModified = blog.lastModified;
     if (!!blog.tags) {
       this.tags = blog.tags;
     } else {
       this.tags = [];
     }
+    this.isDraft = blog.isDraft === true ? true : false;
+  }
+
+  get createdDateDisplay(): string {
+    return dayjs(this.createdDate).format('MMM DD, YYYY');
+  }
+
+  get lastModifiedDisplay(): string {
+    return dayjs(this.lastModified).format('MMM DD, YYYY');
   }
 }
